+++ ui/src/app/core/settings.service.ts
@@ -35,6 +35,8 @@ interface EnvInterface {
   setupWizardComplete: boolean
   recommendChildBridges: boolean
   usePnpm: boolean
+  scheduledBackupDisable: boolean
+  scheduledBackupPath: string
 }
 
 interface AppSettingsInterface {